fix: move check require address function to expense utils and refactor to make it more readable
