+++ lib/utils.js
@@ -1,9 +1,6 @@
 import loadScript from 'load-script';
 import sanitizeHtml from 'sanitize-html';
 
-import { CollectiveType } from './constants/collectives';
-import expenseTypes from './constants/expenseTypes';
-
 export function truncate(str, length) {
   if (!str || typeof str !== 'string' || str.length <= length) {
     return str;
@@ -218,16 +215,3 @@ export const flattenObjectDeep = obj =>
   );
 
 export const stripHTML = htmlContent => sanitizeHtml(htmlContent, { allowedTags: [], allowedAttributes: {} });
-
-export const checkRequiresAddress = values => {
-  const isSelfHostedCollective = values.payee.id === values.payee.host?.id;
-  return (
-    values.payee &&
-    (![CollectiveType.COLLECTIVE, CollectiveType.EVENT, CollectiveType.PROJECT, CollectiveType.FUND].includes(
-      values.payee.type,
-    ) ||
-      isSelfHostedCollective) &&
-    !values.payee.isInvite &&
-    [expenseTypes.INVOICE, expenseTypes.FUNDING_REQUEST, expenseTypes.GRANT].includes(values.type)
-  );
-};
